+++ package_monitor/__init__.py
@@ -2,7 +2,8 @@
 and outstanding updates for Alliance Auth.
 """
 
+# pylint: disable=invalid-name
 default_app_config = "package_monitor.apps.PackageMonitorConfig"
 
-__version__ = "1.11.0"
+__version__ = "1.11.1"
 __title__ = "Package Monitor"