+++ package_monitor/managers.py
@@ -1,3 +1,5 @@
+"""Managers for Package Monitor."""
+
 from typing import Dict, Set
 
 from django.db import models
@@ -27,7 +29,10 @@ logger = LoggerAddTag(get_extension_logger(__name__), __title__)
 
 
 class DistributionQuerySet(models.QuerySet):
+    """QuerySets for Distribution."""
+
     def outdated_count(self) -> int:
+        """Return count of outdated packages."""
         return self.filter(is_outdated=True).count()
 
     def build_install_command(self) -> str:
@@ -60,16 +65,16 @@ class DistributionQuerySet(models.QuerySet):
 
 
 class DistributionManagerBase(models.Manager):
-    def update_all(
-        self, use_threads: bool = False, notifications_disabled: bool = False
-    ) -> int:
+    """Manager for Distribution."""
+
+    def update_all(self, notifications_disabled: bool = False) -> int:
         """Update the list of relevant distribution packages in the database."""
         logger.info(
             f"Started refreshing approx. {self.count()} distribution packages..."
         )
         packages = gather_distribution_packages()
         requirements = compile_package_requirements(packages)
-        update_packages_from_pypi(packages, requirements, use_threads)
+        update_packages_from_pypi(packages, requirements)
         self._save_packages(
             packages=packages,
             requirements=requirements,