+++ package_monitor/tasks.py
@@ -1,3 +1,5 @@
+"""Tasks for Package Monitor."""
+
 from celery import shared_task
 
 from allianceauth.services.hooks import get_extension_logger
@@ -11,4 +13,5 @@ logger = LoggerAddTag(get_extension_logger(__name__), __title__)
 
 @shared_task(time_limit=3600)
 def update_distributions():
+    """Update or create all distributions."""
     Distribution.objects.update_all()