+++ package_monitor/tests/core/test_distribution_packages.py
@@ -1,8 +1,9 @@
 from collections import namedtuple
-from unittest import mock
+from unittest import IsolatedAsyncioTestCase, mock
 
+import aiohttp
+from aioresponses import aioresponses
 from packaging.specifiers import SpecifierSet
-from requests_mock import Mocker
 
 from app_utils.testing import NoSocketsTestCase
 
@@ -127,19 +128,19 @@ class TestCompilePackageRequirements(NoSocketsTestCase):
         expected = {"alpha": {"bravo": SpecifierSet(">=1.0.0")}}
         self.assertDictEqual(expected, result)
 
-    # def test_should_ignore_python_version_requirements(self):
-    #     # given
-    #     dist_alpha = DistributionPackageFactory(name="alpha")
-    #     dist_bravo = DistributionPackageFactory(name="bravo", requires=["alpha>=1.0.0"])
-    #     dist_charlie = DistributionPackageFactory(
-    #         name="charlie", requires=["alpha >= 1.0.0 ; python_version < 3.7"]
-    #     )
-    #     packages = make_packages(dist_alpha, dist_bravo, dist_charlie)
-    #     # when
-    #     result = compile_package_requirements(packages)
-    #     # then
-    #     expected = {"alpha": {"bravo": SpecifierSet(">=1.0.0")}}
-    #     self.assertDictEqual(expected, result)
+    def test_should_ignore_python_version_requirements(self):
+        # given
+        dist_alpha = DistributionPackageFactory(name="alpha")
+        dist_bravo = DistributionPackageFactory(name="bravo", requires=["alpha>=1.0.0"])
+        dist_charlie = DistributionPackageFactory(
+            name="charlie", requires=['alpha >= 1.0.0 ; python_version < "3.7"']
+        )
+        packages = make_packages(dist_alpha, dist_bravo, dist_charlie)
+        # when
+        result = compile_package_requirements(packages)
+        # then
+        expected = {"alpha": {"bravo": SpecifierSet(">=1.0.0")}}
+        self.assertDictEqual(expected, result)
 
     def test_should_ignore_invalid_extra_requirements(self):
         # given
@@ -155,200 +156,190 @@ class TestCompilePackageRequirements(NoSocketsTestCase):
         expected = {"alpha": {"bravo": SpecifierSet(">=1.0.0")}}
         self.assertDictEqual(expected, result)
 
-    # TODO: This test breaks with packaging<22, which is currently required by Auth.
 
-    # def test_should_ignore_invalid_python_release_spec(self, requests_mocker):
-    #     # given
-    #     dist_alpha = DistributionPackageFactory(name="alpha", current="1.0.0")
-    #     packages = make_packages(dist_alpha)
-    #     requirements = {}
-    #     pypi_alpha = PypiFactory(distribution=dist_alpha)
-    #     pypi_alpha.releases["1.1.0"] = [PypiReleaseFactory(requires_python=">=3.4.*")]
-    #     requests_mocker.register_uri(
-    #         "GET", "https://pypi.org/pypi/alpha/json", json=pypi_alpha.asdict()
-    #     )
-    #     # when
-    #     update_packages_from_pypi(
-    #         packages=packages, requirements=requirements, use_threads=False
-    #     )
-    #     # then
-    #     self.assertEqual(packages["alpha"].latest, "1.0.0")
-
-
-@Mocker()
-class TestUpdatePackagesFromPyPi(NoSocketsTestCase):
-    def test_should_update_packages(self, requests_mocker):
+@mock.patch(MODULE_PATH + ".DistributionPackage._fetch_data_from_pypi_async")
+class TestUpdatePackagesFromPyPi(IsolatedAsyncioTestCase):
+    async def test_should_update_packages(self, mock_fetch_data_from_pypi_async):
         # given
         dist_alpha = DistributionPackageFactory(name="alpha", current="1.0.0")
         requirements = {}
         pypi_alpha = PypiFactory(distribution=dist_alpha)
         pypi_alpha.releases["1.1.0"] = [PypiReleaseFactory()]
-        requests_mocker.register_uri(
-            "GET", "https://pypi.org/pypi/alpha/json", json=pypi_alpha.asdict()
-        )
+        mock_fetch_data_from_pypi_async.return_value = pypi_alpha.asdict()
         # when
-        dist_alpha.update_from_pypi(requirements=requirements)
+        await dist_alpha.update_from_pypi_async(
+            requirements=requirements, session=mock.MagicMock()
+        )
         # then
         self.assertEqual(dist_alpha.latest, "1.1.0")
         self.assertEqual(dist_alpha.homepage_url, "https://pypi.org/project/alpha/")
 
-    def test_should_ignore_prereleases_when_stable(self, requests_mocker):
+    async def test_should_ignore_prereleases_when_stable(
+        self, mock_fetch_data_from_pypi_async
+    ):
         # given
         dist_alpha = DistributionPackageFactory(name="alpha", current="1.0.0")
         requirements = {}
         pypi_alpha = PypiFactory(distribution=dist_alpha)
         pypi_alpha.releases["1.1.0a1"] = [PypiReleaseFactory()]
-        requests_mocker.register_uri(
-            "GET", "https://pypi.org/pypi/alpha/json", json=pypi_alpha.asdict()
-        )
+        mock_fetch_data_from_pypi_async.return_value = pypi_alpha.asdict()
         # when
-        dist_alpha.update_from_pypi(requirements=requirements)
+        await dist_alpha.update_from_pypi_async(
+            requirements=requirements, session=mock.MagicMock()
+        )
+
         # then
         self.assertEqual(dist_alpha.latest, "1.0.0")
 
-    def test_should_include_prereleases_when_prerelease(self, requests_mocker):
+    async def test_should_include_prereleases_when_prerelease(
+        self, mock_fetch_data_from_pypi_async
+    ):
         # given
         dist_alpha = DistributionPackageFactory(name="alpha", current="1.0.0a1")
         requirements = {}
         pypi_alpha = PypiFactory(distribution=dist_alpha)
         pypi_alpha.releases["1.0.0a2"] = [PypiReleaseFactory()]
-        requests_mocker.register_uri(
-            "GET", "https://pypi.org/pypi/alpha/json", json=pypi_alpha.asdict()
-        )
+        mock_fetch_data_from_pypi_async.return_value = pypi_alpha.asdict()
         # when
-        dist_alpha.update_from_pypi(requirements=requirements)
+        await dist_alpha.update_from_pypi_async(
+            requirements=requirements, session=mock.MagicMock()
+        )
+
         # then
         self.assertEqual(dist_alpha.latest, "1.0.0a2")
 
-    def test_should_not_update_package_on_network_error(self, requests_mocker):
+    async def test_should_not_update_package_on_network_error(
+        self, mock_fetch_data_from_pypi_async
+    ):
         # given
         dist_alpha = DistributionPackageFactory(name="alpha", current="1.0.0")
         requirements = {}
         pypi_alpha = PypiFactory(distribution=dist_alpha)
         pypi_alpha.releases["1.1.0"] = [PypiReleaseFactory()]
-        requests_mocker.register_uri(
-            "GET",
-            "https://pypi.org/pypi/alpha/json",
-            status_code=500,
-            reason="Test error",
-        )
+        mock_fetch_data_from_pypi_async.return_value = None
         # when
-        dist_alpha.update_from_pypi(requirements=requirements)
+        await dist_alpha.update_from_pypi_async(
+            requirements=requirements, session=mock.MagicMock()
+        )
         # then
         self.assertEqual(dist_alpha.latest, "")
 
-    def test_should_ignore_yanked_releases(self, requests_mocker):
+    async def test_should_ignore_yanked_releases(self, mock_fetch_data_from_pypi_async):
         # given
         dist_alpha = DistributionPackageFactory(name="alpha", current="1.0.0")
         requirements = {}
         pypi_alpha = PypiFactory(distribution=dist_alpha)
         pypi_alpha.releases["1.1.0"] = [PypiReleaseFactory(yanked=True)]
-        requests_mocker.register_uri(
-            "GET", "https://pypi.org/pypi/alpha/json", json=pypi_alpha.asdict()
-        )
+        mock_fetch_data_from_pypi_async.return_value = pypi_alpha.asdict()
         # when
-        dist_alpha.update_from_pypi(requirements=requirements)
+        await dist_alpha.update_from_pypi_async(
+            requirements=requirements, session=mock.MagicMock()
+        )
         # then
         self.assertEqual(dist_alpha.latest, "1.0.0")
 
-    @mock.patch(MODULE_PATH + ".sys")
-    def test_should_ignore_releases_with_incompatible_python_requirement(
-        self, requests_mocker, mock_sys
+    async def test_should_ignore_releases_with_incompatible_python_requirement(
+        self, mock_fetch_data_from_pypi_async
     ):
         # given
-        mock_sys.version_info = SysVersionInfo(3, 6, 9)
         dist_alpha = DistributionPackageFactory(name="alpha", current="1.0.0")
         requirements = {}
         pypi_alpha = PypiFactory(distribution=dist_alpha)
         pypi_alpha.releases["1.1.0"] = [PypiReleaseFactory(requires_python=">=3.7")]
-        requests_mocker.register_uri(
-            "GET", "https://pypi.org/pypi/alpha/json", json=pypi_alpha.asdict()
-        )
+        mock_fetch_data_from_pypi_async.return_value = pypi_alpha.asdict()
         # when
-        dist_alpha.update_from_pypi(requirements=requirements)
+        with mock.patch(MODULE_PATH + ".sys") as mock_sys:
+            mock_sys.version_info = SysVersionInfo(3, 6, 9)
+            await dist_alpha.update_from_pypi_async(
+                requirements=requirements, session=mock.MagicMock()
+            )
         # then
         self.assertEqual(dist_alpha.latest, "1.0.0")
 
-    def test_should_ignore_invalid_release_version(self, requests_mocker):
+    async def test_should_ignore_invalid_release_version(
+        self, mock_fetch_data_from_pypi_async
+    ):
         # given
         dist_alpha = DistributionPackageFactory(name="alpha", current="1.0.0")
         requirements = {}
         pypi_alpha = PypiFactory(distribution=dist_alpha)
         pypi_alpha.releases["a3"] = [PypiReleaseFactory()]
-        requests_mocker.register_uri(
-            "GET", "https://pypi.org/pypi/alpha/json", json=pypi_alpha.asdict()
-        )
+        mock_fetch_data_from_pypi_async.return_value = pypi_alpha.asdict()
         # when
-        dist_alpha.update_from_pypi(requirements=requirements)
+        await dist_alpha.update_from_pypi_async(
+            requirements=requirements, session=mock.MagicMock()
+        )
         # then
         self.assertEqual(dist_alpha.latest, "1.0.0")
 
-    def test_should_ignore_releases_not_matching_consolidated_requirements(
-        self, requests_mocker
+    async def test_should_ignore_releases_not_matching_consolidated_requirements(
+        self, mock_fetch_data_from_pypi_async
     ):
         # given
         dist_alpha = DistributionPackageFactory(name="alpha", current="1.0.0")
         requirements = {"alpha": {"bravo": SpecifierSet("<=1.0.0")}}
         pypi_alpha = PypiFactory(distribution=dist_alpha)
         pypi_alpha.releases["1.1.0"] = [PypiReleaseFactory()]
-        requests_mocker.register_uri(
-            "GET", "https://pypi.org/pypi/alpha/json", json=pypi_alpha.asdict()
-        )
+        mock_fetch_data_from_pypi_async.return_value = pypi_alpha.asdict()
         # when
-        dist_alpha.update_from_pypi(requirements=requirements)
+        await dist_alpha.update_from_pypi_async(
+            requirements=requirements, session=mock.MagicMock()
+        )
         # then
         self.assertEqual(dist_alpha.latest, "1.0.0")
 
-    # def test_should_ignore_newest_release_when_its_requirements_are_not_matching(
-    #     self, requests_mocker
-    # ):
-    #     # given
-    #     dist_alpha = DistributionPackageFactory(name="alpha", current="1.0.0")
-    #     requirements = {"bravo": {"charlie": SpecifierSet("<=1.0.0")}}
-    #     pypi_alpha = PypiFactory(distribution=dist_alpha)
-    #     pypi_alpha.releases["1.1.0"] = [PypiReleaseFactory()]
-    #     requests_mocker.register_uri(
-    #         "GET", "https://pypi.org/pypi/alpha/json", json=pypi_alpha.asdict()
-    #     )
-    #     # when
-    #     dist_alpha.update_from_pypi(requirements=requirements)
-    #     # then
-    #     self.assertEqual(dist_alpha.latest, "1.0.0")
+    async def test_should_ignore_invalid_python_release_spec(
+        self, mock_fetch_data_from_pypi_async
+    ):
+        # given
+        dist_alpha = DistributionPackageFactory(name="alpha", current="1.0.0")
+        packages = make_packages(dist_alpha)
+        requirements = {}
+        pypi_alpha = PypiFactory(distribution=dist_alpha)
+        pypi_alpha.releases["1.1.0"] = [PypiReleaseFactory(requires_python=">=3.4.*")]
+        mock_fetch_data_from_pypi_async.return_value = pypi_alpha.asdict()
+        # when
+        await dist_alpha.update_from_pypi_async(
+            requirements=requirements, session=mock.MagicMock()
+        )
+        # then
+        self.assertEqual(packages["alpha"].latest, "1.0.0")
 
 
-@Mocker()
-class TestFetchDataFromPypi(NoSocketsTestCase):
-    def test_should_return_data(self, requests_mocker: Mocker):
+class TestFetchDataFromPypi(IsolatedAsyncioTestCase):
+    @aioresponses()
+    async def test_should_return_data(self, requests_mocker: aioresponses):
         # given
         obj = DistributionPackageFactory(name="alpha")
-        requests_mocker.register_uri(
-            "GET", "https://pypi.org/pypi/alpha/json", json={"alpha": 1}
-        )
+        requests_mocker.get("https://pypi.org/pypi/alpha/json", payload={"alpha": 1})
         # when
-        result = obj._fetch_data_from_pypi()
+        async with aiohttp.ClientSession() as session:
+            result = await obj._fetch_data_from_pypi_async(session)
         # then
         self.assertEqual(result, {"alpha": 1})
 
-    def test_should_return_none_when_package_does_not_exist(
-        self, requests_mocker: Mocker
+    @aioresponses()
+    async def test_should_return_none_when_package_does_not_exist(
+        self, requests_mocker: aioresponses
     ):
         # given
         obj = DistributionPackageFactory(name="alpha")
-        requests_mocker.register_uri(
-            "GET", "https://pypi.org/pypi/alpha/json", status_code=404
-        )
+        requests_mocker.get("https://pypi.org/pypi/alpha/json", status=404)
         # when
-        result = obj._fetch_data_from_pypi()
+        async with aiohttp.ClientSession() as session:
+            result = await obj._fetch_data_from_pypi_async(session)
         # then
         self.assertIsNone(result)
 
-    def test_should_return_none_on_other_http_errors(self, requests_mocker: Mocker):
+    @aioresponses()
+    async def test_should_return_none_on_other_http_errors(
+        self, requests_mocker: aioresponses
+    ):
         # given
         obj = DistributionPackageFactory(name="alpha")
-        requests_mocker.register_uri(
-            "GET", "https://pypi.org/pypi/alpha/json", status_code=500
-        )
+        requests_mocker.get("https://pypi.org/pypi/alpha/json", status=500)
         # when
-        result = obj._fetch_data_from_pypi()
+        async with aiohttp.ClientSession() as session:
+            result = await obj._fetch_data_from_pypi_async(session)
         # then
         self.assertIsNone(result)